+++ jellyfin_kodi/monitor.py
@@ -9,11 +9,10 @@ import threading
 
 from kodi_six import xbmc
 
-import connect
 import player
 from client import get_device_id
 from objects import PlaylistWorker, on_play, on_update, special_listener
-from helper import translate, settings, window, dialog, api, JSONRPC
+from helper import settings, dialog, JSONRPC
 from helper.utils import JsonDebugPrinter
 from jellyfin import Jellyfin
 from helper import LazyLogger
